feat(main): override maps to main
