+++ app/src/main/AndroidManifest.xml
@@ -31,10 +31,8 @@
                 android:value="@string/google_maps_key"/>
 
         <activity
-                android:name=".MapsActivity"
+                android:name=".MainActivity"
                 android:label="@string/title_activity_maps">
-        </activity>
-        <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
